Refactor: move non-infrastructure tests to different namespace.
